# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ERP Pro API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Locations
...
  /suppliers:
    get:
      operationId: listSuppliers
      tags: [suppliers]
      summary: List all suppliers
      responses:
        "200":
          description: Suppliers list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Supplier"
    post:
      operationId: createSupplier
      tags: [suppliers]
      summary: Create a supplier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSupplierRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
  /suppliers/{id}:
    patch:
      operationId: updateSupplier
      tags: [suppliers]
      summary: Update a supplier
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSupplierRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
    delete:
      operationId: deleteSupplier
      tags: [suppliers]
      summary: Delete a supplier
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Sales
  /sales:
    get:
      operationId: listSales
      tags: [sales]
      summary: List all sales
      parameters:
        - in: query
          name: locationId
          schema:
            type: integer
        - in: query
          name: customerId
          schema:
            type: integer
        - in: query
          name: accountId
          schema:
            type: integer
      responses:
        "200":
          description: Sales list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sale"
    post:
      operationId: createSale
      tags: [sales]
      summary: Create a POS sale
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSaleRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
  /sales/{id}:
    get:
      operationId: getSale
      tags: [sales]
      summary: Get a sale by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sale
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"

  # Purchases
...
  /wallets:
    get:
      operationId: listWallets
      tags: [wallets]
      summary: List all wallets
      responses:
        "200":
          description: Wallets list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Wallet"
    post:
      operationId: createWallet
      tags: [wallets]
      summary: Create a wallet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWalletRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallet"
  /wallets/transfer:
    post:
      operationId: walletTransfer
      tags: [wallets]
      summary: Transfer between wallets or accounts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WalletTransferRequest"
      responses:
        "200":
          description: Transfer result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletTransferResult"

  # Audit
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        name:
          type: string
        role:
          type: string
        locationId:
          type: ["integer", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
      required: [id, username, name, role, isActive, createdAt]
    CreateUserRequest:
      type: object
      properties:
        username:
          type: string
        name:
          type: string
        password:
          type: string
        role:
          type: string
        locationId:
          type: ["integer", "null"]
      required: [username, name, password, role]
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
        locationId:
          type: ["integer", "null"]
        isActive:
          type: boolean
        password:
          type: string

    # Location
...
    Supplier:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        balance:
          type: string
        createdAt:
          type: string
      required: [id, name, balance, createdAt]
    CreateSupplierRequest:
      type: object
      properties:
        name:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
      required: [name]
    UpdateSupplierRequest:
      type: object
      properties:
        name:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        address:
          type: ["string", "null"]

    # Sale item
    SaleItem:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        qty:
          type: integer
        unitPrice:
          type: string
        total:
          type: string
      required: [productId, productName, qty, unitPrice, total]

    # Sale
    Sale:
      type: object
      properties:
        id:
          type: integer
        invoiceNo:
          type: string
        customerId:
          type: ["integer", "null"]
        customerName:
          type: ["string", "null"]
        locationId:
          type: ["integer", "null"]
        locationName:
          type: ["string", "null"]
        accountId:
          type: ["integer", "null"]
        accountName:
          type: ["string", "null"]
        userId:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/SaleItem"
        subtotal:
          type: string
        discount:
          type: string
        tax:
          type: string
        total:
          type: string
        amountPaid:
          type: string
        change:
          type: string
        paymentMethod:
          type: string
        status:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, invoiceNo, userId, items, subtotal, discount, tax, total, amountPaid, change, paymentMethod, status, createdAt]
    CreateSaleRequest:
      type: object
      properties:
        customerId:
          type: ["integer", "null"]
        locationId:
          type: ["integer", "null"]
        accountId:
          type: ["integer", "null"]
        userId:
          type: integer
        items:
          type: array
          items:
            type: object
            properties:
              productId:
                type: integer
              qty:
                type: integer
              unitPrice:
                type: string
            required: [productId, qty, unitPrice]
        discount:
          type: string
        tax:
          type: string
        amountPaid:
          type: string
        paymentMethod:
          type: string
        notes:
          type: ["string", "null"]
      required: [userId, items, discount, tax, amountPaid, paymentMethod]

    # Purchase item
...
    Wallet:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
        balance:
          type: string
        currency:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
      required: [id, name, type, balance, currency, isActive, createdAt]
    CreateWalletRequest:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        balance:
          type: string
        currency:
          type: string
      required: [name, type, balance, currency]
    WalletTransferRequest:
      type: object
      properties:
        fromWalletId:
          type: ["integer", "null"]
        toWalletId:
          type: ["integer", "null"]
        fromAccountId:
          type: ["integer", "null"]
        toAccountId:
          type: ["integer", "null"]
        amount:
          type: string
        notes:
          type: ["string", "null"]
      required: [amount]
    WalletTransferResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success, message]

    # Audit
...